+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/computercraft/peripherals/DisplayPedestalPeripheral.kt
@@ -72,4 +72,22 @@ class DisplayPedestalPeripheral(private val blockEntity: DisplayPedestalBlockEnt
     fun setItemRendered(value: Boolean) {
         blockEntity.renderItem = value
     }
+
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (other !is DisplayPedestalPeripheral) return false
+        if (!super.equals(other)) return false
+
+        if (blockEntity != other.blockEntity) return false
+
+        return true
+    }
+
+    override fun hashCode(): Int {
+        var result = super.hashCode()
+        result = 31 * result + blockEntity.hashCode()
+        return result
+    }
+
+
 }