+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/computercraft/peripherals/PeripheralProxyPeripheral.kt
@@ -18,4 +18,22 @@ class PeripheralProxyPeripheral(private val blockEntity: PeripheralProxyBlockEnt
     override fun getAdditionalTypes(): Set<String> {
         return setOf("peripheral_hub")
     }
+
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (other !is PeripheralProxyPeripheral) return false
+        if (!super.equals(other)) return false
+
+        if (blockEntity != other.blockEntity) return false
+
+        return true
+    }
+
+    override fun hashCode(): Int {
+        var result = super.hashCode()
+        result = 31 * result + blockEntity.hashCode()
+        return result
+    }
+
+
 }