+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/computercraft/peripherals/RealityForgerPeripheral.kt
@@ -165,4 +165,26 @@ class RealityForgerPeripheral(
         })
         return MethodResult.of(true)
     }
+
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (other !is RealityForgerPeripheral) return false
+        if (!super.equals(other)) return false
+
+        if (isEnabled != other.isEnabled) return false
+        if (interactionRadius != other.interactionRadius) return false
+        if (peripheralOwner != other.peripheralOwner) return false
+
+        return true
+    }
+
+    override fun hashCode(): Int {
+        var result = super.hashCode()
+        result = 31 * result + isEnabled.hashCode()
+        result = 31 * result + interactionRadius
+        result = 31 * result + peripheralOwner.hashCode()
+        return result
+    }
+
+
 }