+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/computercraft/peripherals/RemoteObserverPeripheral.kt
@@ -54,4 +54,22 @@ class RemoteObserverPeripheral(
             )
         }
     }
+
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (other !is RemoteObserverPeripheral) return false
+        if (!super.equals(other)) return false
+
+        if (blockEntity != other.blockEntity) return false
+
+        return true
+    }
+
+    override fun hashCode(): Int {
+        var result = super.hashCode()
+        result = 31 * result + blockEntity.hashCode()
+        return result
+    }
+
+
 }