+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.2] - 2023-07-27
+
+### Fixed
+
+- Player disabled from entity link
+- Peripheral equals check
+
 ## [1.2.1] - 2023-07-25
 
 ### Added